+++ src/mainwindow.cpp
@@ -636,6 +636,11 @@ void MainWindow::initStyling() {
 
     // there is no system tray in OS X
     ui->actionShow_system_tray->setText(tr("Show menu bar item"));
+
+    // add a padding for the note tag frame so the `add tag` button doesn't
+    // stick to the right corner
+    ui->noteTagFrame->setStyleSheet(
+            "QFrame {border: none; padding-right: 4px;}");
 #endif
 
     // move the note view scrollbar when the note edit scrollbar was moved